Updated Librarian
Important Notes:
Assumed some methods exists in some classes. All the method explanations are done in details in comments.

Summary:

In Library Class:
1-) demandBook method
2-) bookAmount method
3-) changeStock method

In StandartReader Class
1-) borrowBook method
2-) isBorrowed method
3-) returnTheBook method

should be implemented.

Note:
I have assumed that some methods return the index of some elements inside some data sturctures. This can be changed if the data structure wont have indexes or it wont be efficient. I did not know which data structure we are gonna use so just assumed ArrayList for all.