Changes a few parameters
Detailed Explanation of methods required:

In Library Class:
1-) demandBook(Book/Publication demandedBook)
Assumed that there is a method called demandBook in library class that adds the book to the demanded books list.

2-) bookAmount(String bookName, Language bookLanguage)
Assumed that there is a method called bookAmount in library class that returns the amount of a book which has the given language in the library stock

3-)changeStock(Book/Publication theBook, int changeAmount)
Assumed that there is a changeStock method in library class that adds or removes book from the library stock.
the parameter returnedBook is the book that will be added or removed
the parameter 1 is how much it will be added or removed
Ex: 1 means 1 book will be added. -1 means 1 book will be removed
If the book does not exist in the stock and it is trying to be removed throws an exception ? or does not do anything ?

4-)isInStock(String bookName, Language bookLanguage)
Assumed that there is an isInStock method in library class that returns -1 if the book is not found and the index number if it is found.

In StandartReader Class
1-)borrowBook(Book borrowedBook)
Also assumed borrowBook method of reader adds books to the reader's borrowedBook field

2-)isBorrowed(String bookName, Lang bookLanguage)
Assumed that there is a isBorrowed method in reader class that returns the index of the book in borrowedBooks field Returns -1 if not found.

3-) returnTheBook(int indexNo)
Assumed returnTheBook removes the book in the given index from borrowedBooks field of the customer and adds it to pastReadBooks field.

Notes:
-Preferred to use strings when searching for a book, preferred to use book/publication classes when adding/removing books from somewhere.
-The methods which require indexes or return indexes can be changed depending on the data type.