Update Manager.java
In Library Class

1-)Assumed that there is a method named addLibrarian(Librarian addedLibrarian) inside library class
This method adds the librarian to the librarians data field of library class
If the librarian already exists, throws an exception ?

2-)Assumed that there is a method names removeLibrarian(Librarian removedLibrarian) inside library class
This method removes the librarian from the librarians data field of library class.
If the librarian doesnt exists throws an exception ?

3-)Assumed that there is a method named addJanitor(Janitor addedJanitor) inside library class
This method adds the janitor to the janitors data field of library class
If the janitor already exists, throws an exception ?

4-)Assumed that there is a method names removeJanitor(Janitor removedJanitor) inside library class
This method removes the janitor from the janitors data field of library class.
If the janitor doesnt exists throws an exception ?

5-)Assumed that there is a getJanitor(int searchedJID) method inside library class.
This method returns the searched janitor.
Returns null if the janitor does not exist or throws exception ?

6-)Assumed that there is a demandedBooks data field in the library

7-)Assumed there is a addUpcomingEvent(Event addedEvent) method in library class
This method should return true after adding to upcomingEvenets data field.
It should return false if the event is already in the upcoming events list.

8-)Assumed there is a removeEvent(Event removedEvent) method in library class
This method should return true after removing from offeredEvents data field.
It should return false if the event is not in the offeredEvents list.

9-)Assumed there is a offeredEvents data field in library class.

10-)Assumed there is a endEvent(Event endedEvent) method in library class.
This method should remove the endedEvent from upcomingEvents and move it to pastEvents data field of library class.
Returns the removed event.
If the event is not found in upcomingEvents data field returns null.

11-)Assumed that there is isLibrarian(int librarianID) method in library class
This method returns the index of the librarian inside the librarians data field in library class
If the librarian does not exist, returns -1

12-)Assumed that there is isJanitor(int JanitorID) method in library class
This method returns the index of the janitor inside the janitors data field in library class
If the janitor does not exist, returns -1

In Janitor Class

1-)Assumed that there is addTask(Task newTask) method in janitor class.
This method adds the task to the janitor, returns true.
If the janitor already has the task returns false.

Note: I left some of the methods empty but it's fine for now